pyproject.toml
[project]
name = "document-knowledge-hub"
version = "0.1.0"
description = "A Document Knowledge Hub built with FastAPI"
authors = [
    { name = "Purnendu Kar", email = "purnendu@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12,<4.0"

dependencies = [
    "fastapi>=0.111.0,<1.0.0",
    "uvicorn[standard]>=0.30.0,<1.0.0",
    "sqlalchemy>=2.0.29,<3.0.0",
    "alembic>=1.13.1,<2.0.0",
    "pydantic>=2.7.0,<3.0.0",
    "python-multipart>=0.0.9,<1.0.0",
    "PyPDF2>=3.0.1,<4.0.0",
    "python-docx>=0.8.11,<1.0.0",
    "bcrypt>=4.0.0,<4.1.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "PyJWT>=2.8.0,<3.0.0",
    "slowapi>=0.1.9,<1.0.0",
    "itsdangerous>=2.2.0,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
